Updating calories to be an Integer

+++ backend/src/main/java/com/mealchemy/engine/dto/NutritionRequest.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record NutritionRequest(
-    @JsonProperty("calories_kcal") BigDecimal caloriesKcal,
+    @JsonProperty("calories_kcal") Integer caloriesKcal,
     @JsonProperty("protein_g") BigDecimal proteinG,
     @JsonProperty("carbs_g") BigDecimal carbsG,
     @JsonProperty("fat_g") BigDecimal fatG
